+++ WoWsShipBuilder.Core/DataUI/Constants.cs
@@ -3,5 +3,6 @@ namespace WoWsShipBuilder.Core.DataUI
     public static class Constants
     {
         public const decimal TickRate = 0.142857142857m;
+        public const decimal KnotsToMps = 2.6854m;
     }
 }