+++ ReleaseNotes.md
@@ -5,13 +5,15 @@
 - Added an indicator to show if there is any talent or conditional modifier active
 - Added Hungarian translation. Thanks to StaK_1980!
 - Added amount of HP restored per heal on ship consumables.
+- Added shell trajectory chart.
 
 ## Changes
-- Increase spacing between rows in the ship statistic for better readability.
+- Increased spacing between rows in the ship statistic for better readability.
 
 ## Bugfixes
 - Fixed Hidden Menace and Damage Control Party Modification 1 not changing stats of damage control party reload for CV
 - Fixed some localization issues.
+- Fixed incorrect values for CV torpedo speed and arming distance.
 ___
 To download the program for the first time, you need to download only the Setup.exe file. The other files are used by the application for automatic updates.